chore: add optional dependency suppport for expo CLI and improved docs
