+++ docusaurus/docs/reactnative/customization/native_handlers.mdx
@@ -41,6 +41,14 @@ An `async` function that compresses an image and returns the local `uri` of the 
 
 **Expo:** [`expo-image-manipulator`](https://docs.expo.io/versions/latest/sdk/imagemanipulator/)
 
+### `setClipboardString`
+
+An function to copy strings or text in the message.
+
+**React Native:** [`react-native-clipboard`](https://github.com/react-native-clipboard/clipboard)
+
+**Expo:** [`expo-clipboard`](https://docs.expo.dev/versions/v48.0.0/sdk/clipboard/)
+
 ### `deleteFile`
 
 A function that deletes a file at a given local `uri`.
