+++ examples/ExpoMessaging/package.json
@@ -15,14 +15,10 @@
     "expo": "~47.0.9",
     "expo-av": "~13.0.2",
     "expo-blur": "~12.0.1",
-    "expo-clipboard": "~4.0.1",
-    "expo-document-picker": "~11.0.1",
     "expo-file-system": "~15.1.1",
-    "expo-haptics": "~12.0.1",
     "expo-image-manipulator": "~11.0.0",
     "expo-image-picker": "~14.0.2",
     "expo-media-library": "~15.0.0",
-    "expo-sharing": "~11.0.1",
     "expo-splash-screen": "~0.17.5",
     "expo-status-bar": "~1.4.2",
     "react": "18.1.0",
