+++ package/expo-package/src/handlers/Sound.ts
@@ -1,14 +1,8 @@
-import type { AVPlaybackSource, AVPlaybackStatus, AVPlaybackStatusToSet } from 'expo-av';
-
 import { AudioComponent } from '../optionalDependencies/Video';
 
 export const Sound = {
   initializeSound: AudioComponent
-    ? async (
-        source: AVPlaybackSource,
-        initialStatus: AVPlaybackStatusToSet,
-        onPlaybackStatusUpdate: (playbackStatus: AVPlaybackStatus) => void,
-      ) => {
+    ? async (source, initialStatus, onPlaybackStatusUpdate: (playbackStatus) => void) => {
         const { sound } = await AudioComponent.Sound.createAsync(
           source,
           initialStatus,
