+++ package/expo-package/src/handlers/index.ts
@@ -3,12 +3,8 @@ export * from './deleteFile';
 export * from './getLocalAssetUri';
 export * from './getPhotos';
 export * from './NetInfo';
-export * from './pickDocument';
 export * from './saveFile';
-export * from './setClipboardString';
-export * from './shareImage';
 export * from './Sound';
 export * from './takePhoto';
-export * from './triggerHaptic';
 export * from './Video';
 export * from './oniOS14GalleryLibrarySelectionChange';
