--- package/expo-package/src/handlers/setClipboardString.ts
@@ -1,10 +1,0 @@
-let Clipboard: { setString: (string: string) => void } | undefined;
-
-try {
-  Clipboard = require('expo-clipboard');
-} catch (e) {
-  // do nothing
-}
-
-export const setClipboardString =
-  Clipboard !== undefined ? (string: string) => Clipboard.setString(string) : null;
