+++ package/expo-package/src/index.js
@@ -9,15 +9,18 @@ import {
   getPhotos,
   NetInfo,
   oniOS14GalleryLibrarySelectionChange,
-  pickDocument,
   saveFile,
-  setClipboardString,
-  shareImage,
   Sound,
   takePhoto,
-  triggerHaptic,
   Video,
 } from './handlers';
+
+import {
+  pickDocument,
+  setClipboardString,
+  shareImage,
+  triggerHaptic,
+} from './optionalDependencies';
 
 registerNativeHandlers({
   compressImage,
