+++ package/expo-package/src/optionalDependencies/index.ts
@@ -1,0 +1,5 @@
+export * from './setClipboardString';
+export * from './shareImage';
+export * from './pickDocument';
+export * from './triggerHaptic';
+export * from './Video';
