+++ package/expo-package/src/optionalDependencies/setClipboardString.ts
@@ -1,10 +1,10 @@
 let Clipboard: { setString: (string: string) => void } | undefined;
 
 try {
-  Clipboard = require('@react-native-clipboard/clipboard').default;
+  Clipboard = require('expo-clipboard').default;
 } catch (e) {
   // do nothing
-  console.log('@react-native-clipboard/clipboard is not installed');
+  console.log('expo-clipboard is not installed');
 }
 
 export const setClipboardString = Clipboard
